requirements.txt
streamlit
pandas
scikit-learn
transformers
sentence-transformers
faiss-cpu
joblib
spacy
